Split out BukuCrypt class into its own file
